requirements.txt

pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
scipy==1.10.1
beautifulsoup4==4.12.2
requests==2.28.2
python-dotenv==1.0.0